remove duplicate nav parse function
