+++ src/demos/main_gsdc2021.py
@@ -5,12 +5,13 @@ import gzip
 import argparse
 import sys
 
+import prx.util
 from prx import helpers, main
 
 log = helpers.get_logger(__name__)
 
 
-@helpers.timeit
+@prx.util.timeit
 def process_gsdc2021_dataset(path_dataset: Path):
     """
     This script creates a local copy of gunzipped RINEX files from the GSDC2021 data set and process them with prx
