+++ src/prx/rinex_nav/nav_file_discovery.py
@@ -6,6 +6,7 @@ from pathlib import Path
 import georinex
 import urllib.request
 import pandas as pd
+import prx.util
 import requests
 
 from prx import converters, helpers
@@ -40,7 +41,7 @@ def try_downloading_ephemerides_http(day: pd.Timestamp, local_destination_folder
         local_file = converters.compressed_to_uncompressed(local_compressed_file)
         os.remove(local_compressed_file)
         log.info(f"Downloaded broadcast ephemerides file from {url}")
-        helpers.repair_with_gfzrnx(local_file)
+        prx.util.repair_with_gfzrnx(local_file)
         return local_file
     except Exception as e:
         log.warning(f"Could not download broadcast ephemerides file for {day}: {e}")
@@ -83,7 +84,7 @@ def try_downloading_ephemerides_ftp(day: pd.Timestamp, folder: Path):
     local_file = converters.compressed_to_uncompressed(local_compressed_file)
     os.remove(local_compressed_file)
     log.info(f"Downloaded broadcast ephemerides file {ftp_file}")
-    helpers.repair_with_gfzrnx(local_file)
+    prx.util.repair_with_gfzrnx(local_file)
     return local_file
